Move init command in its own folder
