+++ lib/index.js
@@ -1,4 +1,4 @@
-const initCommand = require('./init.command')
+const initCommand = require('./init/init.command')
 const runCommand = require('./run.command')
 const deployCommand = require('./deploy.command')
