+++ lib/init/init-stop-by-user.error.js
