+++ lib/init/init.command.js
@@ -2,9 +2,9 @@ const fs = require('fs')
 const path = require('path')
 const inquirer = require('inquirer')
 
-const { FsUtils } = require('./shared/fs.utils')
-const { log } = require('./shared/log.utils')
-const { gitClone } = require('./shared/git.utils')
+const { FsUtils } = require('../shared/fs.utils')
+const { log } = require('../shared/log.utils')
+const { gitClone } = require('../shared/git.utils')
 
 const { InitStopedByUser } = require('./init-stop-by-user.error')
